consensus/kaiabft: move speculative execution into backend
